pyproject.toml
[project]
name = "mps-fallback-bench"
version = "0.1.0"
description = "Benchmark CPU fallback costs for PyTorch MPS ops"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "requests>=2.31"
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy", "matplotlib"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["bench", "ops", "detect", "metrics", "report", "report.templates"]

